[layout] : Account page publishing
